Compare the returning device's rate in the integer domain it was stored in
-Wfloat-equal, GCC-only and so the strict job's alone to find: the recovery
check compared the device's double sample rate against granted.sampleRate, which
is an int. `open` fills that field with static_cast<int> of the same
getCurrentSampleRate(), so casting the same way here makes the round trip exact
by construction rather than approximately true, and there is no float comparison
left to be unsafe.

+++ src/wfg/engine/audio/DeviceLayer.cpp
@@ -156,7 +156,9 @@ namespace wfg::audio
             observedBlock.store (device->getCurrentBufferSizeSamples(), std::memory_order_relaxed);
             recovery.started (device->getTypeName() == pinnedType
                 && device->getName() == pinnedSetup.outputDeviceName
-                && device->getCurrentSampleRate() == granted.sampleRate
+                // The int domain `granted.sampleRate` was stored in, by the same
+                // cast (`open` below), so the round trip compares exactly.
+                && static_cast<int> (device->getCurrentSampleRate()) == granted.sampleRate
                 && device->getCurrentBufferSizeSamples() == granted.blockSize
                 && device->getActiveInputChannels() == pinnedSetup.inputChannels
                 && device->getActiveOutputChannels() == pinnedSetup.outputChannels);
